Revised all data models based off actual data from Census. Converted all configs to dists and excluding real files. Corrected SolTech ID.

+++ app/models/events/achievementEarned.model.js
@@ -23,11 +23,11 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.INTEGER.UNSIGNED
         },
         world_id: {
-            type: Sequelize.INTEGER.UNSIGNED
+            type: Sequelize.SMALLINT.UNSIGNED
         },
         zone_id: {
-            type: Sequelize.INTEGER.UNSIGNED
-        }
+            type: Sequelize.MEDIUMINT.UNSIGNED
+        },
     });
   
     return AchievementEarned;
