+++ app/models/events/continentUnlock.model.js
@@ -23,10 +23,10 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.STRING
         },
         world_id: {
-            type: Sequelize.TINYINT.UNSIGNED
+            type: Sequelize.SMALLINT.UNSIGNED
         },
         zone_id: {
-            type: Sequelize.TINYINT.UNSIGNED
+            type: Sequelize.MEDIUMINT.UNSIGNED
         },
         triggering_faction: {
             type: Sequelize.TINYINT.UNSIGNED
