+++ app/models/events/metagameEvent.model.js
@@ -41,11 +41,11 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.TINYINT.UNSIGNED
         },
         world_id: {
-            type: Sequelize.TINYINT.UNSIGNED
+            type: Sequelize.SMALLINT.UNSIGNED
         },
         zone_id: {
-            type: Sequelize.TINYINT.UNSIGNED
-        }
+            type: Sequelize.MEDIUMINT.UNSIGNED
+        },
     });
   
     return MetagameEvent;
