+++ app/models/events/playerLogin.model.js
@@ -17,8 +17,8 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.STRING
         },
         world_id: {
-            type: Sequelize.TINYINT.UNSIGNED
-        }
+            type: Sequelize.SMALLINT.UNSIGNED
+        },
     });
   
     return PlayerLogin;
